Update gitignore, add toml, move logs
