+++ icarus_venv/bin/numpy-config
@@ -1,0 +1,8 @@
+#!/home/fitchs/Projects/Icarus_v2/icarus_venv/bin/python3.12
+# -*- coding: utf-8 -*-
+import re
+import sys
+from numpy._configtool import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
