+++ icarus_venv/bin/pyside6-qmlimportscanner
@@ -1,0 +1,8 @@
+#!/home/fitchs/Projects/Icarus_v2/icarus_venv/bin/python3.12
+# -*- coding: utf-8 -*-
+import re
+import sys
+from PySide6.scripts.pyside_tool import qmlimportscanner
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(qmlimportscanner())
