+++ icarus_venv/bin/python
@@ -1,0 +1,1 @@
+python3.12
