+++ icarus_venv/pyvenv.cfg
@@ -1,0 +1,5 @@
+home = /usr/bin
+include-system-site-packages = false
+version = 3.12.3
+executable = /usr/bin/python3.12
+command = /usr/bin/python3.12 -m venv /home/fitchs/Projects/Icarus_v2/icarus_venv
